+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ADDED:
 
 - new bean class for hold information of a certificate extension
 - new factory class for create KeyPairGenerator objects
+- new extension class ProviderExtensions for provide security provider information from the current jdk
 
 CHANGED: